chore: move trace convertor to traces package
